added helper function for transaction statistics process
